+++ src/main/java/nl/dead_pixel/telebot/api/types/files/Document.java
@@ -15,7 +15,7 @@ public class Document implements IFile {
     @JsonProperty("mime_type")
     private String mimeType;
     @JsonProperty("file_size")
-    private int fileSize;
+    private Long fileSize;
 
     /**
      * Gets file id.
@@ -102,7 +102,7 @@ public class Document implements IFile {
      *
      * @return the file size
      */
-    public int getFileSize() {
+    public Long getFileSize() {
         return fileSize;
     }
 
@@ -112,7 +112,7 @@ public class Document implements IFile {
      * @param fileSize the file size
      * @return the file size
      */
-    public Document setFileSize(int fileSize) {
+    public Document setFileSize(Long fileSize) {
         this.fileSize = fileSize;
         return this;
     }
